requirements.txt
Django==5.0.6
django-environ==0.11.2
black==24.4.2
pillow==10.3.0
djangorestframework==3.15.1
pytest==8.2.2
pytest-django==4.8.0